Refactor function resolution logic into FunctionResolver
